requirements.txt
requests
pyyaml
tk